Refactor plugin system: unify plugin type definitions and reintroduce Compartment class
