+++ packages/react/src/index.ts
@@ -95,6 +95,5 @@ export function wrapper(draggableFactory: ReturnType<typeof createDraggable>) {
 }
 
 export const useDraggable = wrapper(draggable_factory);
-export { Compartment } from '@neodrag/core';
 export * from '@neodrag/core/plugins';
 export const instances = draggable_factory.instances;
